salvando ajustes

+++ src/SME.SGP.Aplicacao/Commands/Ocorrencia/AlterarOcorrencia/AlterarOcorrenciaCommand.cs
@@ -19,7 +19,7 @@ namespace SME.SGP.Aplicacao
         public bool ConsideraHistorico { get; set; }
         public int AnoLetivo { get; set; }
         public int DreId { get; set; }
-        public int UeId { get; set; }
+        public long UeId { get; set; }
         public int Modalidade { get; set; }
         public int Semestre { get; set; }
         public long TurmaId { get; set; }
